+++ modules/src/address.rs
@@ -3,18 +3,18 @@
 use std::convert::TryFrom;
 
 use anomaly::{BoxError, Context};
-use bech32::FromBase32;
 use bech32::ToBase32;
+use bech32::{FromBase32, Variant};
 
 use tendermint::account::Id as AccountId;
 
 pub fn account_to_string(addr: AccountId) -> Result<String, BoxError> {
-    Ok(bech32::encode("cosmos", addr.to_base32())
+    Ok(bech32::encode("cosmos", addr.to_base32(), Variant::Bech32)
         .map_err(|e| Context::new("cannot generate bech32 account", Some(e.into())))?)
 }
 
 pub fn string_to_account(raw: String) -> Result<AccountId, BoxError> {
-    let (_hrp, data) =
+    let (_hrp, data, _variant) =
         bech32::decode(&raw).map_err(|e| Context::new("bad signer", Some(e.into())))?;
     let addr_bytes =
         Vec::<u8>::from_base32(&data).map_err(|e| Context::new("bad signer", Some(e.into())))?;